section of getting articles by category completed

+++ article/article_service.go
@@ -1,8 +1,11 @@
 package article
 
 import (
+	"bitroom/constants"
 	article_model "bitroom/models/article"
 	"bitroom/types"
+	"bitroom/utils"
+	"net/http"
 )
 
 type ArticleService struct {
@@ -18,6 +21,14 @@ func NewArticleService(store ArticleStoreInterface) *ArticleService {
 // --------------------------------------------------------------------------------------------------------------------
 
 func (a *ArticleService) AddArticle(data *NewArticle) (*Article, *types.CustomError) {
+	// check category exists
+	exists, existsErr := a.store.CheckCategoryExist(data.Category)
+	if existsErr != nil {
+		return nil, existsErr
+	}
+	if !exists {
+		return nil, utils.NewError(constants.NotFound, http.StatusNotFound)
+	}
 	article, err := a.store.AddArticle(data)
 	if err != nil {
 		return nil, err
@@ -27,7 +38,7 @@ func (a *ArticleService) AddArticle(data *NewArticle) (*Article, *types.CustomEr
 
 // --------------------------------------------------------------------------------------------------------------------
 
-func (a *ArticleService) GetArticles() ([]MinimumArtilce, *types.CustomError) {
+func (a *ArticleService) GetArticles() ([]MinimumArticle, *types.CustomError) {
 	articles, err := a.store.GetArticles()
 	if err != nil {
 		return nil, err
@@ -38,18 +49,40 @@ func (a *ArticleService) GetArticles() ([]MinimumArtilce, *types.CustomError) {
 // --------------------------------------------------------------------------------------------------------------------
 
 func (a *ArticleService) GetArticleById(id uint) (*article_model.Article, *types.CustomError) {
+	// check article exist
+	exists, checkingErr := a.store.CheckArticleExist(id)
+	if checkingErr != nil {
+		return nil, checkingErr
+	}
+	if !exists {
+		return nil, utils.NewError(constants.NotFound, http.StatusNotFound)
+	}
+	// get article
 	article, err := a.store.GetArticleById(id)
 	if err != nil {
 		return nil, err
 	}
+	// success
 	return article, nil
 }
 
 // --------------------------------------------------------------------------------------------------------------------
 
-func (a *ArticleService) GetArticlesByCategory(categoryId uint) ([]MinimumArtilce, *types.CustomError) {
-
-	return nil, nil
+func (a *ArticleService) GetArticlesByCategory(categoryId uint) ([]MinimumArticle, *types.CustomError) {
+	// check category exist
+	exist, existsErr := a.store.CheckCategoryExist(categoryId)
+	if existsErr != nil {
+		return nil, existsErr
+	}
+	if !exist {
+		return nil, utils.NewError(constants.NotFound, http.StatusNotFound)
+	}
+	// get articles
+	artilces, err := a.store.GetArticlesByCategory(categoryId)
+	if err != nil {
+		return nil, err
+	}
+	return artilces, nil
 }
 
 // --------------------------------------------------------------------------------------------------------------------
@@ -62,16 +95,26 @@ func (a *ArticleService) EditArticle() (*Article, *types.CustomError) {
 // --------------------------------------------------------------------------------------------------------------------
 
 func (a *ArticleService) DeleteArticle(id uint) *types.CustomError {
+	// check article exist
+	exists, checkingErr := a.store.CheckArticleExist(id)
+	if checkingErr != nil {
+		return checkingErr
+	}
+	if !exists {
+		return utils.NewError(constants.NotFound, http.StatusNotFound)
+	}
+	// delete article
 	err := a.store.DeleteArticle(id)
 	if err != nil {
 		return err
 	}
+	// success
 	return nil
 }
 
 // --------------------------------------------------------------------------------------------------------------------
 
-func (a *ArticleService) GetPopularArticles() ([]MinimumArtilce, *types.CustomError) {
+func (a *ArticleService) GetPopularArticles() ([]MinimumArticle, *types.CustomError) {
 
 	return nil, nil
 }
